Improved workaround for uris for imported datasets
Remove uris for datasets and resources that were generated by test environment(s).

If a dataset doesn't have a uri after removing the test-environment one, and we have to generate one on exporting to XML, then use the permalink. (We can't use the dataset id because we need to strip out the test-environment uri when importing the dataset, and at that point it doesn't have an id.)

If a resource doesn't have a uri after removing the test-environment one, then use its generated url on this environment. If we're importing the dataset, then we can't do that because neither dataset nor resource have an id yet. In that case, just leave the resource uri empty. It will get generated on export to XML.