fix: stop harvester in gather in case of a parsing error on any page
Case: in case of harvesting data catalogs that use hydra in
order to serve the datasets in pages

Problem: when not all pages can be parsed, we want to prevent
the harvester assuming that the datasets that have not been
parsed have been deleted at the origin. Because then these
datasets would be deleted on opendata.swiss as well.

Solution: stop the harvester at the gather stage in case
any of the pages does not have content or can not be parsed
In this case a gather error will be reported and the hydra
information is also rendered with this error, so that the
publisher knows on what page the error occured.

Because of this fix it is no longer necessary to overwrite
the gather stage for an addtional error reporting.