feat: Check temporal dates match the given datatype
The aim here is to reject poorly formatted dates/datetimes, e.g. a value of 2020-01-01 given with a datatype of xsd:dateTime. We handle dates/datetime values differently depending on the datatype, and  check that the format matches before calculating the end datetime.

This approach cannot fully work, however, because when we calculate the dataset graph using rdflib, all literals with a datatype of xsd:date, xsd:gYearMonth or xsd:gYear are parsed as dates using isoformat. xsd:dateTime literals are parsed as datetimes using isoformat. This means that an xsd:gYear value of '2000' is passed to us as '2000-01-01', and we can't check whether it was originally given as '2000' or, e.g., '2000-4'. The latter value is incorrectly formatted for any datatype. isoformat will interpret both values as year=2000, no month or day given, and returns us the first day of that year.

We get the unicode representation of the date or datetime literal out of the graph, and only take the values we expect to get from it: an xsd:gYear value is trimmed down to just the year before being processed (getting the datetime of the end of that year), and an xsd:gYearMonth is trimmed to just the year and month. This can prevent some errors, but not all.

Notably, xsd:dateTime values that don't include a time, a day or a month value will still be parsed as datetimes.