fix: Update check for values that are nested in a list
For each language in the multilang dict, we might get a string or a list of values, e.g. {'de': 'Dataset Title'} or {'de': ['tag1', 'tag2', 'tag3']}. If it's a string, we need to wrap it in a list so we can safely iterate over it in the next line - otherwise we iterate by letter: 'D', 'a', 't', ...

Checking whether the value had the '__iter__' attribute worked for this in Python 2, but in Python 3, strings apparently also have this attribute. Let's just check whether the value is a list or not. If there are any other cases, we can catch them later!